refactor src/database-server --> src/data-model ref #3
