+++ src/data-model/CMakeLists.txt
@@ -5,7 +5,6 @@ if(NOT MAKE_EXE)
     message(SEND_ERROR "Make executable is required to build sqlite3.")
 endif(NOT MAKE_EXE)
 
-include(ExternalProject)
 ExternalProject_Add(sqlite3_project
     SOURCE_DIR ${CMAKE_SOURCE_DIR}/3rdparty/sqlite3
     BINARY_DIR ${CMAKE_CURRENT_BINARY_DIR}/sqlite3
@@ -39,26 +38,6 @@ if(BUILD_TESTING)
   add_library(test_util ${CMAKE_CURRENT_SOURCE_DIR}/test/testutil.cc)
   target_include_directories(test_util SYSTEM PUBLIC ${Boost_INCLUDE_DIR})
   target_link_libraries(test_util PUBLIC ${Boost_LIBRARIES})
-  include(GoogleTest)
-  ExternalProject_Add(googletest_project
-      SOURCE_DIR ${CMAKE_SOURCE_DIR}/3rdparty/gtest
-      BINARY_DIR ${CMAKE_CURRENT_BINARY_DIR}/gtest
-      URL https://github.com/google/googletest/archive/release-1.8.1.tar.gz
-      INSTALL_COMMAND echo "Skipping install"
-  )
-  ExternalProject_Get_Property(googletest_project source_dir)
-  set(GTEST_INCLUDE_DIR ${source_dir}/googletest/include)
-  ExternalProject_Get_Property(googletest_project binary_dir)
-  set(GTEST_LIBRARY gtest)
-  set(GTEST_LIBRARY_PATH ${binary_dir}/googlemock/gtest/${CMAKE_FIND_LIBRARY_PREFIXES}${GTEST_LIBRARY}.a)
-  add_library(${GTEST_LIBRARY} STATIC IMPORTED)
-  set_property(TARGET ${GTEST_LIBRARY} PROPERTY IMPORTED_LOCATION ${GTEST_LIBRARY_PATH} )
-  add_dependencies(${GTEST_LIBRARY} googletest_project)
-  set(GTEST_MAIN_LIBRARY ${GTEST_LIBRARY}_main)
-  set(GTEST_MAIN_LIBRARY_PATH ${binary_dir}/googlemock/gtest/${CMAKE_FIND_LIBRARY_PREFIXES}${GTEST_MAIN_LIBRARY}.a)
-  add_library(${GTEST_MAIN_LIBRARY} STATIC IMPORTED)
-  set_property(TARGET ${GTEST_MAIN_LIBRARY} PROPERTY IMPORTED_LOCATION ${GTEST_MAIN_LIBRARY_PATH})
-  add_dependencies(${GTEST_MAIN_LIBRARY} googletest_project)
   add_executable(sqlitedatamodel_test
     ${CMAKE_CURRENT_SOURCE_DIR}/test/sqlitedatamodel_integration_test.cc)
   target_include_directories(sqlitedatamodel_test SYSTEM PRIVATE ${GTEST_INCLUDE_DIR})
