+++ src/data-model/test/testutil.cc
@@ -36,7 +36,7 @@
     [DAQs-URL]: https://github.com/denisacostaq/DAQs "DAQs"
  */
 
-#include "src/database-server/test/testutil.h"
+#include "src/data-model/test/testutil.h"
 
 #include <boost/filesystem.hpp>
